fix compilation

+++ pcap/pcap_linux.go
@@ -696,8 +696,8 @@ func openOfflineFile(file *os.File) (handle *Handle, err error) {
 
 func (b *BPF) pcapOfflineFilter(ci gopacket.CaptureInfo, data []byte) bool {
 	hdr := (*C.struct_pcap_pkthdr)(&b.hdr)
-	hdr.ts.tv_sec = C.int32_t(C.gopacket_time_secs_t(ci.Timestamp.Unix()))
-	hdr.ts.tv_usec = C.int32_t(C.gopacket_time_usecs_t(ci.Timestamp.Nanosecond() / 1000))
+	hdr.ts.tv_sec = C.int64_t(C.gopacket_time_secs_t(ci.Timestamp.Unix()))
+	hdr.ts.tv_usec = C.int64_t(C.gopacket_time_usecs_t(ci.Timestamp.Nanosecond() / 1000))
 	hdr.caplen = C.bpf_u_int32(len(data)) // Trust actual length over ci.Length.
 	hdr.len = C.bpf_u_int32(ci.Length)
 	dataptr := (*C.u_char)(unsafe.Pointer(&data[0]))
